+++ packages/resources/assets/BaseSite/custom-resource/s3-upload.py
@@ -56,13 +56,16 @@ def s3_deploy(s3_source_zip, s3_dest, file_options, replace_values):
         logger.info("| replacing pattern: %s", pattern)
         for filepath in glob.iglob(pattern, recursive=True):
             logger.info("| replacing pattern in file %s", filepath)
-            with open(filepath) as file:
-                ori = file.read()
-                new = ori.replace(replace_value['search'], replace_value['replace'])
-                if ori != new:
-                    logger.info("| updated")
-                    with open(filepath, "w") as file:
-                        file.write(new)
+            # In Gatsby, you can have folders named like a file
+            # ie. 404.html/ and it will get matched by "**/*.html"
+            if os.path.isfile(filepath):
+                with open(filepath) as file:
+                    ori = file.read()
+                    new = ori.replace(replace_value['search'], replace_value['replace'])
+                    if ori != new:
+                        logger.info("| updated")
+                        with open(filepath, "w") as file:
+                            file.write(new)
 
     # sync from "contents" to destination
     for file_option in file_options: