+++ packages/resources/src/Auth.ts
@@ -22,7 +22,9 @@ export interface ApiAttachmentProps {
    * @example
    * ```js
    * const api = new Api(stack, "Api", {});
-   * const auth = new Auth(stack, "Auth")
+   * const auth = new Auth(stack, "Auth", {
+   *   authenticator: "functions/authenticator.handler"
+   * })
    * auth.attach({
    *   api
    * })
@@ -38,7 +40,9 @@ export interface ApiAttachmentProps {
    * @example
    * ```js
    * const api = new Api(stack, "Api", {});
-   * const auth = new Auth(stack, "Auth")
+   * const auth = new Auth(stack, "Auth", {
+   *   authenticator: "functions/authenticator.handler"
+   * })
    * auth.attach({
    *   api,
    *   prefix: "/custom/prefix"
@@ -59,7 +63,13 @@ export interface ApiAttachmentProps {
  * })
  */
 export class Auth extends Construct {
+  /**
+   * Secret that contains the public JWT signing key
+   */
   public readonly SST_AUTH_PUBLIC: Secret;
+  /**
+   * Secret that contains the private JWT signing key
+   */
   public readonly SST_AUTH_PRIVATE: Secret;
   private readonly authenticator: FunctionDefinition;
 
@@ -105,7 +115,7 @@ export class Auth extends Construct {
     const policyStatement = new PolicyStatement({
       actions: ["ssm:PutParameter", "ssm:DeleteParameter"],
       effect: Effect.ALLOW,
-      resources: ["*"]
+      resources: ["*"],
     });
     stack.customResourceHandler.addToRolePolicy(policyStatement);
 
@@ -114,11 +124,25 @@ export class Auth extends Construct {
       resourceType: "Custom::AuthKeys",
       properties: {
         publicPath,
-        privatePath
-      }
+        privatePath,
+      },
     });
   }
 
+  /**
+   * Attaches auth construct to an API
+   *
+   * @example
+   * ```js
+   * const api = new Api(stack, "Api", {});
+   * const auth = new Auth(stack, "Auth", {
+   *   authenticator: "functions/authenticator.handler"
+   * })
+   * auth.attach({
+   *   api
+   * })
+   * ```
+   */
   public attach(scope: Construct, props: ApiAttachmentProps) {
     if (this.apis.has(props.api))
       throw new Error(
@@ -130,8 +154,8 @@ export class Auth extends Construct {
       props.api.addRoutes(scope, {
         [path]: {
           type: "function",
-          function: this.authenticator
-        }
+          function: this.authenticator,
+        },
       });
       props.api.getFunction(path)!.addConfig([this.SST_AUTH_PRIVATE]);
       props.api.getFunction(path)!.addEnvironment("SST_AUTH_PREFIX", prefix);
@@ -150,6 +174,9 @@ export class Auth extends Construct {
       }
     }
   }
+  /**
+   * @internal
+   */
   public static injectConfig() {
     for (const auth of Auth.list) {
       auth.injectConfig();