+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @serverless-stack/core
 
+## 1.15.7
+
+## 1.15.6
+
+## 1.15.5
+
+## 1.15.4
+
+### Patch Changes
+
+- [`3f241696a`](https://github.com/serverless-stack/sst/commit/3f241696a0b0d24d8a582fb037267e82ca0d809b) Thanks [@thdxr](https://github.com/thdxr)! - Support camelcase for kysely type generation
+
 ## 1.15.3
 
 ### Patch Changes